rename json.xpect.tests to json.xpect.ui.tests
