--- tests/org.eclipse.n4js.json.xpect.tests/xpect/nonui/validation/DuplicateKeys.json.xt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2016 NumberFour AG.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   NumberFour AG - Initial API and implementation
- */
-
-/* XPECT_SETUP org.eclipse.n4js.json.tests.JSONXpectTest END_SETUP  */
-{
-	"a": 1,
-	"b": 2,
-	// Duplicate key is detected by validation XPECT warnings --> "Property a duplicates property a (line 14)." at ""a""
-	"a": "str",
-	"nested": {
-		// Differently nested key is not detected as duplicate XPECT noerrors
-		"a" : "not a duplicate",
-		// Differently nested key is not detected as duplicate XPECT noerrors
-		"b" : ["not a duplicate", 12]
-	}
-}
