+++ tests/org.eclipse.n4js.json.xpect.ui.tests/.settings/de.loskutov.anyedit.AnyEditTools.prefs
