+++ tests/org.eclipse.n4js.json.xpect.ui.tests/.settings/org.eclipse.pde.prefs
