+++ tests/org.eclipse.n4js.json.xpect.ui.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,29 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %pluginName
+Bundle-SymbolicName: org.eclipse.n4js.json.xpect.ui.tests
+Bundle-Version: 0.0.1.qualifier
+Fragment-Host: org.eclipse.n4js.json;bundle-version="0.0.1"
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.eclipse.xtext;visibility:=reexport,
+ org.eclipse.xtext.generator;resolution:=optional,
+ org.apache.commons.logging;bundle-version="1.0.4";resolution:=optional,
+ org.eclipse.emf.codegen.ecore;resolution:=optional,
+ org.eclipse.emf.mwe.utils;resolution:=optional,
+ org.eclipse.emf.mwe2.launch;resolution:=optional,
+ org.eclipse.xtext.util,
+ org.eclipse.emf.ecore,
+ org.eclipse.emf.common,
+ org.antlr.runtime,
+ org.eclipse.core.resources,
+ org.eclipse.core.runtime,
+ org.eclipse.xtext.testing,
+ org.eclipse.n4js.json.tests.helper,
+ org.eclipse.xpect.xtext.lib;bundle-version="0.2.0",
+ org.eclipse.n4js.utils,
+ org.eclipse.n4js.tests.helper,
+ org.eclipse.n4js.xpect,
+ org.eclipse.n4js.json.ui,
+ org.eclipse.n4js.xpect.ui
+Bundle-Vendor: %providerName
+Automatic-Module-Name: org.eclipse.n4js.json.xpect.ui.tests
