+++ tests/org.eclipse.n4js.json.xpect.ui.tests/build.properties
