+++ tests/org.eclipse.n4js.json.xpect.ui.tests/xpect/nonui/formatter/FormatJSONNull.json.xt
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2016 NumberFour AG.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   NumberFour AG - Initial API and implementation
+ */
+
+/* XPECT_SETUP org.eclipse.n4js.json.xpect.ui.tests.JSONXpectPluginTest END_SETUP  */
+/* XPECT formattedLines 1 ---
+null
+--- */
+null
