+++ tests/org.eclipse.n4js.json.xpect.ui.tests/xpect/nonui/formatter/FormatNestedJSONObject.json.xt
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2016 NumberFour AG.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   NumberFour AG - Initial API and implementation
+ */
+
+/* XPECT_SETUP org.eclipse.n4js.json.xpect.ui.tests.JSONXpectPluginTest END_SETUP  */
+/* XPECT formattedLines 28 ---
+{
+	"a": 1,
+	"b": 2,
+	"f": [
+		1,
+		2,
+		3
+	],
+	"c": [
+		1,
+		2,
+		3,
+		[
+			1,
+			2,
+			3
+		]
+	],
+	"d": {
+		"a": 1,
+		"b": 1,
+		"c": 3,
+		"d": [
+			1,
+			2,
+			3
+		]
+	}
+}
+--- */
+{
+	"a": 1,
+	"b": 2,
+	"f": [1, 2, 3 ],
+	"c": [1, 2, 3, 
+		[
+			1,
+			2,
+			3
+		]
+	],
+	"d": {
+		"a": 1, "b": 1, "c": 3, "d": [ 1, 2, 3]
+	}
+}
