+++ tests/pom.xml
@@ -86,6 +86,8 @@ Contributors:
 
 		<!-- Manifest tests -->
 		<module>org.eclipse.n4js.n4mf.xpect.ui.tests</module>
+		<!-- JSON tests -->
+		<module>org.eclipse.n4js.json.xpect.ui.tests</module>
 		<!-- Runner tests -->
 		<module>org.eclipse.n4js.runner.ui.tests</module>
